Rename workspace search daemon to flashgrep
