--- resources/codgrep/README.md
@@ -1,8 +1,0 @@
-Place the prebuilt `cg` daemon binary in this directory.
-
-Expected filenames:
-
-- macOS/Linux: `cg`
-- Windows: `cg.exe`
-
-BitFun dev/build scripts load the daemon from this repository-relative path.
