+++ resources/flashgrep/README.md
@@ -1,0 +1,8 @@
+Place the prebuilt `flashgrep` daemon binary in this directory.
+
+Expected filenames:
+
+- macOS/Linux: `flashgrep`
+- Windows: `flashgrep.exe`
+
+BitFun dev/build scripts load the daemon from this repository-relative path.
