+++ src/apps/desktop/tauri.conf.json
@@ -19,7 +19,7 @@
     "resources": {
       "../../mobile-web/dist": "mobile-web/dist",
       "resources/worker_host.js": "resources/worker_host.js",
-      "../../../resources/codgrep": "codgrep"
+      "../../../resources/flashgrep": "flashgrep"
     },
     "linux": {
       "deb": {
