+++ src/crates/core/src/agentic/tools/implementations/grep_tool.rs
@@ -674,9 +674,9 @@ mod tests {
             repo_status: WorkspaceSearchRepoStatus {
                 repo_id: "repo".to_string(),
                 repo_path: "/repo".to_string(),
-                storage_root: "/repo/.bitfun/search/codgrep-index".to_string(),
-                base_snapshot_root: "/repo/.bitfun/search/codgrep-index/base-snapshot".to_string(),
-                workspace_overlay_root: "/repo/.bitfun/search/codgrep-index/workspace-overlay"
+                storage_root: "/repo/.bitfun/search/flashgrep-index".to_string(),
+                base_snapshot_root: "/repo/.bitfun/search/flashgrep-index/base-snapshot".to_string(),
+                workspace_overlay_root: "/repo/.bitfun/search/flashgrep-index/workspace-overlay"
                     .to_string(),
                 phase: WorkspaceSearchRepoPhase::Ready,
                 snapshot_key: None,
