--- src/crates/core/src/service/search/codgrep/daemon/mod.rs
@@ -1,4 +1,0 @@
-mod managed;
-pub(crate) mod protocol;
-
-pub(crate) use managed::{ManagedDaemonClient, OpenedRepo};
