--- src/crates/core/src/service/search/codgrep/mod.rs
@@ -1,3 +1,0 @@
-pub mod daemon;
-pub mod error;
-pub mod sdk;
