+++ src/crates/core/src/service/search/flashgrep/daemon/mod.rs
@@ -1,0 +1,4 @@
+mod managed;
+pub(crate) mod protocol;
+
+pub(crate) use managed::{ManagedDaemonClient, OpenedRepo};
