+++ src/crates/core/src/service/search/flashgrep/error.rs
@@ -1,0 +1,13 @@
+use std::io;
+
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum AppError {
+    #[error("io error: {0}")]
+    Io(#[from] io::Error),
+    #[error("protocol error: {0}")]
+    Protocol(String),
+}
+
+pub type Result<T> = std::result::Result<T, AppError>;
