+++ src/crates/core/src/service/search/flashgrep/mod.rs
@@ -1,0 +1,3 @@
+pub mod daemon;
+pub mod error;
+pub mod sdk;
